clean up services tests
